refactor: presentation layer
