+++ src/presentation/stream.rs
