+++ src/services/mod.rs
@@ -1,9 +1,9 @@
 pub mod actor_registry;
 pub mod aof;
-pub mod client;
 pub mod cluster;
 pub mod config;
 pub mod error;
 pub mod interface;
+
 pub mod query_io;
 pub mod statefuls;
